move tests to kalix.javasdk
