+++ sdk/spring-sdk-testkit/src/test/java/kalix/javasdk/testmodels/valueentity/CounterValueEntity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package spring.testmodels.valueentity;
+package kalix.javasdk.testmodels.valueentity;
 
 import kalix.javasdk.valueentity.ValueEntity;
