+++ sdk/spring-sdk-testkit/src/test/java/kalix/javasdk/testmodels/valueentity/CounterValueEntityTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package spring.testmodels.valueentity;
+package kalix.javasdk.testmodels.valueentity;
 
 import kalix.javasdk.testkit.ValueEntityResult;
 import kalix.javasdk.testkit.ValueEntityTestKit;
